@@ -1,139 +1,147 @@
 {
   "name": "@lyrio/ui",
-  "version": "0.3.4",
+  "version": "0.5.6",
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.16.7",
-    "@babel/plugin-proposal-decorators": "^7.17.2",
-    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-decorators": "^7.18.9",
     "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
-    "@esm-bundle/markdown-it": "^12.3.2",
-    "@fontsource/dm-mono": "^4.5.2",
-    "@fontsource/fira-code": "^4.5.5",
-    "@fontsource/inconsolata": "^4.5.2",
-    "@fontsource/jetbrains-mono": "^4.5.3",
-    "@fontsource/noto-sans": "^4.5.3",
-    "@fontsource/noto-serif": "^4.5.3",
-    "@fontsource/nunito-sans": "^4.5.2",
-    "@fontsource/open-sans": "^4.5.5",
-    "@fontsource/pt-mono": "^4.5.1",
-    "@fontsource/pt-sans": "^4.5.1",
-    "@fontsource/roboto": "^4.5.3",
-    "@fontsource/roboto-mono": "^4.5.3",
-    "@fontsource/roboto-slab": "^4.5.4",
-    "@fontsource/source-code-pro": "^4.5.4",
-    "@fontsource/source-serif-pro": "^4.5.2",
-    "@fontsource/ubuntu-mono": "^4.5.2",
-    "@fontsource/zilla-slab": "^4.5.1",
+    "@esm-bundle/markdown-it": "^13.0.1",
+    "@fontsource/dm-mono": "^4.5.8",
+    "@fontsource/fira-code": "^4.5.10",
+    "@fontsource/inconsolata": "^4.5.6",
+    "@fontsource/jetbrains-mono": "^4.5.9",
+    "@fontsource/noto-sans": "^4.5.10",
+    "@fontsource/noto-serif": "^4.5.9",
+    "@fontsource/nunito-sans": "^4.5.8",
+    "@fontsource/open-sans": "^4.5.10",
+    "@fontsource/pt-mono": "^4.5.9",
+    "@fontsource/pt-sans": "^4.5.9",
+    "@fontsource/roboto": "^4.5.7",
+    "@fontsource/roboto-mono": "^4.5.7",
+    "@fontsource/roboto-slab": "^4.5.8",
+    "@fontsource/source-code-pro": "^4.5.10",
+    "@fontsource/source-serif-pro": "^4.5.8",
+    "@fontsource/ubuntu-mono": "^4.5.10",
+    "@fontsource/zilla-slab": "^4.5.8",
     "@hackmd/markdown-it-task-lists": "^2.1.1",
-    "@tanem/react-nprogress": "^4.0.7",
+    "@monaco-editor/react": "^4.4.5",
+    "@tanem/react-nprogress": "^5.0.9",
     "@types/blueimp-md5": "^2.18.0",
-    "@types/chart.js": "^2.9.35",
+    "@types/chart.js": "^2.9.37",
     "@types/css": "^0.0.33",
     "@types/flat": "^5.0.2",
     "@types/gravatar": "^1.8.3",
     "@types/html-minifier-terser": "^6.1.0",
     "@types/js-yaml": "^4.0.5",
-    "@types/lodash": "^4.14.178",
+    "@types/lodash": "^4.14.182",
     "@types/markdown-it": "^12.2.3",
-    "@types/node": "16",
+    "@types/node": "18.6.3",
     "@types/object-path": "^0.11.1",
     "@types/prismjs": "^1.26.0",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/serialize-javascript": "^5.0.2",
     "@types/streamsaver": "^2.0.1",
-    "@types/svgo": "^2.6.2",
-    "@types/twemoji": "^12.1.2",
+    "@types/svgo": "^2.6.3",
+    "@types/twemoji": "^13.1.2",
     "@types/ua-parser-js": "^0.7.36",
     "@types/uuid": "^8.3.4",
-    "@types/validator": "^13.7.1",
-    "@vitejs/plugin-legacy": "^1.7.1",
-    "@vitejs/plugin-react": "^1.2.0",
+    "@types/validator": "^13.7.4",
+    "@vitejs/plugin-legacy": "^2.0.0",
+    "@vitejs/plugin-react": "^2.0.0",
     "ansi_up": "^5.1.0",
     "autotrack": "^2.4.1",
-    "axios": "^0.26.0",
-    "babel-plugin-import": "^1.13.3",
+    "axios": "^0.27.2",
+    "babel-plugin-import": "^1.13.5",
     "babel-plugin-prismjs": "^2.1.0",
     "blueimp-md5": "^2.19.0",
-    "browserslist-generator": "^1.0.65",
-    "chart.js": "^3.7.1",
+    "browserslist-generator": "^1.0.66",
+    "chart.js": "^3.8.2",
     "class-validator": "^0.13.2",
     "css": "^3.0.0",
     "css-element-queries": "^1.2.3",
     "css-url-relative-plugin": "^1.1.0",
     "customize-cra": "^1.0.0",
     "cz-conventional-changelog": "3.3.0",
-    "dayjs": "^1.10.7",
+    "dayjs": "^1.11.4",
     "dtsgenerator": "2.7.0",
+    "esbuild-plugin-browserslist": "^0.4.13",
     "flat": "^5.0.2",
     "fomantic-ui-css": "^2.8.8",
-    "fs-extra": "^10.0.0",
-    "git-repo-info": "^2.1.1",
+    "fs-extra": "^10.1.0",
+    "getscript-promise": "^0.0.1",
     "gravatar": "^1.8.2",
     "hack-font": "source-foundry/Hack#1eddf94",
-    "husky": "^7.0.4",
+    "husky": "^8.0.1",
     "immutability-helper": "^3.1.1",
     "js-yaml": "^4.1.0",
     "jsondiffpatch": "^0.4.1",
-    "less": "^4.1.2",
-    "lodash-es": "^4.17.21",
+    "less": "^4.1.3",
+    "lodash": "^4.17.21",
     "markdown-it-math-loose": "^4.1.1",
     "markdown-it-mentions": "^1.0.0",
-    "markdown-it-merge-cells": "^1.0.1",
-    "mathjax-full": "^3.2.0",
-    "mobx": "^6.3.13",
+    "markdown-it-merge-cells": "^2.0.0",
+    "mathjax-full": "^3.2.2",
+    "mobx": "^6.6.1",
     "mobx-persist": "^0.4.1",
-    "mobx-react-lite": "^3.2.3",
-    "mobx-utils": "^6.0.4",
-    "monaco-ace-tokenizer": "^0.2.1",
-    "monaco-editor": "0.32.1",
-    "monaco-themes": "^0.4.0",
-    "navi": "^0.14.0",
-    "node-html-parser": "^5.2.0",
-    "noty": "^3.2.0-beta-deprecated",
+    "mobx-react-lite": "^3.4.0",
+    "mobx-utils": "^6.0.5",
+    "monaco-ace-tokenizer": "^0.2.2",
+    "monaco-editor": "0.33.0",
+    "monaco-themes": "^0.4.2",
+    "navi": "^0.15.0",
+    "node-html-parser": "^5.3.3",
+    "noty": "3.1.4",
     "object-path": "^0.11.8",
     "p-all": "^4.0.0",
     "patch-package": "^6.4.7",
+    "plausible-tracker": "^0.3.8",
     "postinstall-postinstall": "^2.1.0",
-    "prettier": "^2.5.1",
-    "prismjs": "^1.26.0",
-    "react": "^17.0.2",
-    "react-chartjs-2": "4.0.1",
+    "prettier": "^2.7.1",
+    "prismjs": "^1.28.0",
+    "purgecss": "^4.1.3",
+    "react": "^18.2.0",
+    "react-chartjs-2": "4.3.1",
     "react-countdown": "^2.3.2",
-    "react-dom": "^17.0.2",
-    "react-google-recaptcha-v3": "^1.9.7",
-    "react-helmet-async": "^1.2.2",
-    "react-intl": "^5.24.6",
-    "react-monaco-editor": "0.47.0",
-    "react-navi": "^0.14.4",
-    "react-textarea-autosize": "^8.3.3",
+    "react-dom": "^18.2.0",
+    "react-google-recaptcha-v3": "^1.10.0",
+    "react-helmet-async": "^1.3.0",
+    "react-intl": "^6.0.5",
+    "react-navi": "^0.15.0",
+    "react-textarea-autosize": "^8.3.4",
     "rollup-plugin-node-polyfills": "^0.2.1",
-    "semantic-ui-react": "^2.1.1",
+    "rollup-plugin-visualizer": "^5.7.1",
+    "sass": "^1.54.4",
+    "semantic-ui-react": "^2.1.3",
     "serialize-javascript": "^6.0.0",
+    "serve": "^14.0.1",
     "sf-mono-webfont": "^1.0.0",
-    "socket.io-client": "^4.4.1",
+    "socket.io-client": "^4.5.1",
     "socket.io-msgpack-parser": "^3.0.1",
     "streamsaver": "^2.0.6",
     "string-replace-to-array": "^2.1.0",
     "svgo": "^2.8.0",
+    "terser": "^5.14.2",
     "timeago.js": "^4.0.2",
-    "twemoji": "^13.1.0",
+    "twemoji": "^14.0.2",
     "typeface-lato": "^1.1.13",
     "typeface-saira": "^1.1.13",
-    "typescript": "~4.5.5",
+    "typescript": "~4.7.4",
     "ua-parser-js": "^1.0.2",
-    "use-debounce": "^7.0.1",
+    "use-debounce": "^8.0.3",
     "uuid": "^8.3.2",
     "validator": "^13.7.0",
-    "vite": "^2.8.1",
+    "vite": "^3.0.4",
     "vite-plugin-compile-time": "^0.1.2",
-    "vite-plugin-ejs": "^1.4.3",
-    "vite-plugin-html-minifier-terser": "^2.0.0",
+    "vite-plugin-ejs": "^1.5.0",
+    "vite-plugin-externals": "^0.5.1",
+    "vite-plugin-html-minifier-terser": "^2.0.3",
     "vite-plugin-prismjs": "^0.0.8",
-    "vite-plugin-public-path": "^0.0.3",
+    "vite-plugin-public-path": "^1.0.0",
+    "vite-plugin-purgecss": "^0.2.12",
     "wastyle": "^0.0.5",
-    "web-streams-polyfill": "^3.2.0",
-    "xss": "^1.0.10"
+    "web-streams-polyfill": "^3.2.1",
+    "xss": "^1.0.13"
   },
   "scripts": {
     "start": "vite",
