@@ -1,10 +1,19 @@
 import React, { lazy, Suspense } from "react";
 import { Loader } from "semantic-ui-react";
+import { loader as monacoLoader } from "@monaco-editor/react";
 import type { CodeEditorProps } from "./CodeEditor";
 
 import style from "./CodeEditor.module.less";
 
-const CodeEditor = lazy(() => import("./CodeEditor"));
+const CodeEditor = lazy(async () => {
+  monacoLoader.config({
+    paths: {
+      vs: `${window.cdnjs}/monaco-editor/${EXTERNAL_PACKAGE_VERSION["monaco-editor"]}/min/vs`
+    }
+  });
+  window["Monaco"] = await monacoLoader.init();
+  return import("./CodeEditor");
+});
 
 const LazyCodeEditor: React.FC<CodeEditorProps> = props => {
   const loading = (
