@@ -97,9 +97,11 @@ return {
       version: {
         name: "Version",
         values: {
-          1.3: "1.3",
-          1.4: "1.4",
-          1.5: "1.5"
+          1.5: "1.5",
+          1.6: "1.6",
+          1.7: "1.7",
+          1.8: "1.8",
+          1.9: "1.9"
         }
       },
       platform: {
@@ -132,8 +134,8 @@ return {
         name: "Version",
         values: {
           2.7: "2.7",
-          3.6: "3.6",
-          3.9: "3.9"
+          3.9: "3.9",
+          "3.10": "3.10"
         }
       }
     }
@@ -167,7 +169,8 @@ return {
         name: "Version",
         values: {
           4.2: "4.2",
-          5: "5"
+          5: "5",
+          6: "6"
         }
       },
       optimize: {
@@ -210,7 +213,8 @@ return {
         name: "Version",
         values: {
           7.3: "7.3",
-          8: "8"
+          8: "8",
+          9: "9"
         }
       }
     }
