@@ -22,9 +22,9 @@ return {
   upload: "Upload",
   confirm_override_question: "Be sure to override file(s) below?",
   confirm_override: "Confirm",
-  progress_hashing: "Preparing {progress}%",
   progress_waiting: "Waiting",
   progress_uploading: "Uploading {progress}%",
+  progress_retrying: "Retrying",
   progress_requesting: "Requesting",
   progress_error: "Error",
   progress_cancelled: "Cancelled",
