@@ -22,9 +22,9 @@ return {
   upload: "アップロード",
   confirm_override_question: "以下のファイルを上書きしますか？",
   confirm_override: "上書きする",
-  progress_hashing: "ファイルを処理中 {progress}%",
   progress_waiting: "アップロードを待機中",
   progress_uploading: "アップロード中 {progress}%",
+  progress_retrying: "[TBT] Retrying",
   progress_requesting: "リクエストしています",
   progress_error: "アップロードに失敗しました",
   progress_cancelled: "アップロードをキャンセルしました",
