@@ -22,9 +22,9 @@ return {
   upload: "上传",
   confirm_override_question: "确认要覆盖以下文件吗？",
   confirm_override: "确认",
-  progress_hashing: "准备上传 {progress}%",
   progress_waiting: "等待上传",
   progress_uploading: "正在上传 {progress}%",
+  progress_retrying: "重试中",
   progress_requesting: "正在请求",
   progress_error: "上传出错",
   progress_cancelled: "上传取消",
