@@ -1,43 +1,42 @@
-import { mathjax } from "mathjax-full/js/mathjax";
-import { TeX } from "mathjax-full/js/input/tex";
-import { SVG } from "mathjax-full/js/output/svg";
-import { browserAdaptor } from "mathjax-full/js/adaptors/browserAdaptor";
-import { HTMLHandler } from "mathjax-full/js/handlers/html/HTMLHandler";
-import { SafeHandler } from "mathjax-full/js/ui/safe/SafeHandler";
+import getScript from "getscript-promise";
 
-// Load TeX packages
-import "mathjax-full/js/input/tex/ams/AmsConfiguration.js";
-import "mathjax-full/js/input/tex/boldsymbol/BoldsymbolConfiguration.js";
-import "mathjax-full/js/input/tex/colorv2/ColorV2Configuration.js";
-import "mathjax-full/js/input/tex/html/HtmlConfiguration.js";
-import "mathjax-full/js/input/tex/noundefined/NoUndefinedConfiguration.js";
-import "mathjax-full/js/input/tex/physics/PhysicsConfiguration.js";
+const TEX_PACKAGES = ["ams", "boldsymbol", "colorv2", "html", "noundefined", "physics"];
 
-mathjax.handlers.register(SafeHandler(new HTMLHandler(browserAdaptor())));
+export const loadMathJax = (() => {
+  let promise = (async () => {
+    window.MathJax = {
+      loader: {
+        load: ["input/tex-base", ...TEX_PACKAGES.map(packageName => `[tex]/${packageName}`), "output/chtml", "ui/safe"]
+      },
+      tex: {
+        packages: ["base", ...TEX_PACKAGES]
+      },
+      chtml: {
+        adaptiveCSS: false
+      }
+    };
+    await getScript(`${window.cdnjs}/mathjax/${EXTERNAL_PACKAGE_VERSION["mathjax-full"]}/es5/startup.js`);
+    await window.MathJax.startup.promise;
+    await window.MathJax.startup.document.updateDocument();
+  })();
+  promise.then(() => (promise = null));
 
-const mathDocument = mathjax.document(document, {
-  InputJax: new TeX({
-    packages: {
-      "[+]": ["ams", "boldsymbol", "colorv2", "html", "noundefined", "physics"]
-    }
-  }),
-  OutputJax: new SVG()
-});
-
-// Add CSS styles
-mathDocument.updateDocument();
+  return () => promise;
+})();
 
 export function renderMath(math: string, display: boolean) {
   try {
-    const wrapper = mathDocument.convert(math, {
+    window.MathJax.texReset();
+    const wrapper = window.MathJax.tex2chtml(math, {
       display
-    });
+    }) as HTMLElement;
 
-    const title = document.createElement("title");
-    title.innerText = math;
+    wrapper.title = math;
 
-    const defsElement = wrapper.getElementsByTagName("defs")[0];
-    defsElement.parentNode.insertBefore(title, defsElement);
+    const copy = document.createElement("span");
+    copy.className = "copy";
+    copy.innerText = math;
+    wrapper.appendChild(copy);
 
     return wrapper;
   } catch (e) {
@@ -45,7 +44,7 @@ export function renderMath(math: string, display: boolean) {
 
     const wrapper = document.createElement("mjx-container");
     wrapper.className = "MathJax";
-    wrapper.setAttribute("jax", "SVG");
+    wrapper.setAttribute("jax", "CHTML");
     if (display) wrapper.setAttribute("display", "true");
 
     const message = document.createElement("span");
