@@ -9,7 +9,7 @@ import { appState } from "@/appState";
 import { Link, useLocalizer, useNavigationChecked, useScreenWidthWithin } from "@/utils/hooks";
 import { defineRoute, RouteError } from "@/AppRouter";
 import api from "@/api";
-import LazyMarkdownContent from "@/markdown/LazyMarkdownContent";
+import MarkdownContent from "@/markdown/MarkdownContent";
 import { getDiscussionDisplayTitle, getDiscussionUrl } from "@/pages/discussion/utils";
 import formatDateTime from "@/utils/formatDateTime";
 import { EmojiRenderer } from "@/components/EmojiRenderer";
@@ -76,7 +76,7 @@ let HomePage: React.FC<HomePageProps> = props => {
   const getNotice = () =>
     props.notice && (
       <Segment className={style.segment} color="pink">
-        <LazyMarkdownContent placeholderLines={7} content={props.notice} />
+        <MarkdownContent placeholderLines={7} content={props.notice} />
       </Segment>
     );
 
@@ -345,7 +345,7 @@ let HomePage: React.FC<HomePageProps> = props => {
                   {inMainView ? (
                     <Table.Cell className={style.columnBio}>
                       {appState.serverPreference.misc.renderMarkdownInUserBio ? (
-                        <LazyMarkdownContent content={user.bio} dontUseContentFont />
+                        <MarkdownContent content={user.bio} dontUseContentFont placeholderLines={1} />
                       ) : (
                         <EmojiRenderer>
                           <div>{user.bio}</div>
@@ -431,14 +431,6 @@ HomePage = observer(HomePage);
 
 export default defineRoute(async request => {
   const dataPromise: ReturnType<typeof fetchData> = fetchData();
-
-  // If user bio is need to be rendered, await the renderer to be loaded.
-  if (
-    appState.serverPreference.misc.renderMarkdownInUserBio &&
-    appState.serverPreference.misc.homepageUserListOnMainView
-  ) {
-    await LazyMarkdownContent.load();
-  }
 
   return <HomePage {...await dataPromise} />;
 });
